requirements.txt
pyTelegramBotAPI>=4.14.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pydub>=0.25.1
NRCLex>=3.0.0
textblob>=0.17.1